+++ packages/components/modules/messages/native/RoomsList/ChatCard/ChatCardOptions/type.ts
@@ -1,0 +1,16 @@
+import { RefObject } from 'react'
+
+import { BottomSheetModal } from '@gorhom/bottom-sheet'
+
+export type ChatCardOptionsProps = {
+  bottomDrawerRef: RefObject<BottomSheetModal | undefined>
+  //   handleSheetChanges: (index: number) => void
+  handleSheetChanges: () => void
+  handleArchiveChat: () => void
+  handleMarkAsUnread: () => void
+  handleChatDetails: () => void
+  handleGoToProfile: () => void
+  handleDeleteChat: () => void
+  isArchived: boolean
+  isArchiveMutationInFlight: boolean
+}