+++ packages/components/modules/messages/native/RoomsList/ChatCard/index.tsx
@@ -1,0 +1,10 @@
+import { ChatCard as ChatCardComponent } from './ChatCardComponent'
+
+const ChatCard = ({ roomRef, isArchived }: { roomRef: any | null; isArchived: boolean }) => {
+  if (!roomRef) {
+    return null
+  }
+  return <ChatCardComponent roomRef={roomRef!} isArchived={isArchived} />
+}
+
+export { ChatCard }