+++ packages/design-system/components/native/buttons/FabButton/types.ts
@@ -1,0 +1,8 @@
+import { Ionicons } from '@expo/vector-icons'
+import { type TouchableOpacityProps } from 'react-native'
+
+export interface FabButtonProps extends TouchableOpacityProps {
+  iconName: keyof typeof Ionicons.glyphMap
+  iconSize?: number
+  iconColor?: string
+}