+++ packages/design-system/components/native/inputs/SearchInput/styles.ts
@@ -1,0 +1,14 @@
+import { StyleSheet } from 'react-native'
+
+import { Theme } from '../../../../styles/native'
+
+export const createStyles = (theme: Theme) =>
+  StyleSheet.create({
+    searchInput: {
+      marginTop: 12,
+      marginBottom: 12,
+      backgroundColor: theme.colors.surface.active,
+      outlineColor: theme.colors.surface.active,
+      height: 45,
+    },
+  })